Add models that exist in various places to demonstrate fracturing
The models that are used by Boo, the standard OO client and the Takari OO client have been added. This
demonstrates what happens when we lack a domain model that we publish to the outside world. If OO
is to proliferate there needs to be a standard model like you see in our exported YAML models and
in Boo templates.